improved category structure & management (#72)
* set default calendar on new user registration

* category types

* category form

* wip: categories edit

* wip: categories edit & balance update logic

* wip: transfer feature

* wip: transfer dialog

* wip: expense transfer between calendars

* fixed phpstan errors

* fixed prettier errors

* fixed migrations & unused types

* cleanup

* bump version

+++ backend/src/Controller/Finance/CategoryController.php
@@ -2,10 +2,11 @@
 
 declare(strict_types=1);
 
-namespace App\Controller\Finance\Category;
+namespace App\Controller\Finance;
 
 use App\Controller\AbstractApiController;
 use App\Entity\Category;
+use App\Enum\CategoryType;
 use App\Repository\CategoryRepository;
 use App\Request\Category\CreateCategoryRequest;
 use App\Request\Category\UpdateCategoryRequest;
@@ -23,9 +24,24 @@ class CategoryController extends AbstractApiController
     }
 
     #[Route('', name: 'list', methods: Request::METHOD_GET)]
-    public function list(): JsonResponse
+    public function list(Request $request): JsonResponse
     {
-        return $this->respond($this->categoryRepository->findBy([], ['name' => 'ASC']));
+        $criteria = [];
+
+        if ($request->query->get('userCategoriesOnly', '1')) {
+            $criteria = ['type' => [
+                CategoryType::USER,
+                CategoryType::UNCATEGORIZED,
+            ]];
+        }
+
+        return $this->respond($this->categoryRepository->findBy($criteria, ['name' => 'ASC']));
+    }
+
+    #[Route('/system', name: 'list_system', methods: Request::METHOD_GET)]
+    public function system(): JsonResponse
+    {
+        return $this->respond($this->categoryRepository->findSystem());
     }
 
     #[Route('/{category}', name: 'get', methods: Request::METHOD_GET)]
@@ -40,6 +56,7 @@ class CategoryController extends AbstractApiController
         $category = (new Category())
             ->setName($request->getName())
             ->setColor($request->getColor())
+            ->setType(CategoryType::USER)
         ;
 
         $this->categoryRepository->save($category);
@@ -63,6 +80,10 @@ class CategoryController extends AbstractApiController
     #[Route('/{category}', name: 'remove', methods: Request::METHOD_DELETE)]
     public function remove(Category $category): JsonResponse
     {
+        if (!$category->isDefinedByUser()) {
+            throw $this->createAccessDeniedException();
+        }
+
         $this->categoryRepository->remove($category);
 
         return $this->respond(new EmptyResponse());
